+++ api-derive/chain/getBlockByNumber.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.7';
-import type { AnyNumber } from 'https://deno.land/x/polkadot@0.2.17/types/types/index.ts';
+import type { AnyNumber } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { SignedBlockExtended } from '../type/types.ts';
 import type { DeriveApi } from '../types.ts';