+++ types/interfaces/contracts/runtime.ts
@@ -3,7 +3,7 @@
 
 import type { DefinitionCall, DefinitionsCall } from '../../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.17/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 const SHARED_V1_V2: Record<string, DefinitionCall> = {
   get_storage: {