+++ types/primitive/Data.ts
@@ -1,11 +1,11 @@
 // Copyright 2017-2022 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { IU8a, Registry } from 'https://deno.land/x/polkadot@0.2.17/types-codec/types/index.ts';
+import type { IU8a, Registry } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 import type { H256 } from '../interfaces/runtime/index.ts';
 
-import { Bytes, Enum } from 'https://deno.land/x/polkadot@0.2.17/types-codec/mod.ts';
-import { isString, isU8a, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.17/util/mod.ts';
+import { Bytes, Enum } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import { isString, isU8a, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 /** @internal */
 function decodeDataU8a (registry: Registry, value: Uint8Array): [undefined | Uint8Array, number | undefined] {